2021_11_18_ProjectOnArtificialIntelligence_HandWrittenDigitClassification/requirements.txt
click==8.0.3
cycler==0.11.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.31
threadpoolctl==3.1.0
Werkzeug==2.0.2
